Ondo USDY: fix mantle and cosmos chain balances
The Mantle balance for USDY references an 'unreleased' address pointing
to an unknown multisig. This should be removed, as we consider all
tokens on Mantle to be in circulation.

The Noble and Injective balances have been reworked to use native RPCs
and fetch balances using the IBC channel balances. This is a lot more
scalable, and I was able to add Osmosis as well.